implemented the controls section of the Settings menu
The settings menu functions as follows:
Attatch an s_rebinder component to the button.
Add the actions you wish to rebind.
Call SetKey() OnClick().
The text will change to reflect the rebinding and the key will be rebound.

Limitations:
Cannot happen on the main menu as there is no player to reference. The option is not present on the main menu.
When rebinding multiple actions to a single key, other keys will only add the rebinding, not remove any old ones. Advise only biding one action as it currently stands.
You can rebind the same key to multiple actions.
Not every action has a UI option to rebind.